Bug: "junit.swingui cannot be resolved" error when compiling "net.sf.j2s.test.junit.WorldTest"
The class "WorldTest" referenced "junit.swingui.TestRunner.run" to run
"WorldTest" as a Unit test with a Swing UI.

This is the only reference to "junit.swingui" in the whole project.

As "WorldTest" did not do any real test (beside a "assertTrue(false)")
"WorldTest " was removed, thus also fixing the bug.