Bug: 'The method flatten ... is not applicable for the arguments ...'
Eclipse 4.7.3a reports Java problem:

    'The method flatten(Node<T>, IntFunction<T[]>) in the type Nodes is not
     applicable for the arguments (Node, IntFunction)'

when compiling ReferencePipeline.java.

Cause
=====

It looks like the compiler cannot correctly infer the (generic) types
of the method calls in the argument list of flatten.

Solution
========

Introduce an extra temporary variable 'node' with explicit type 'Node<A>'
holding the value of the expression originally passed to flatten. This
way the correct types are used in the method call and the error message
disappears.