Java2Script 3.2.2 java.lang.reflect.Proxy fix
Note that functional interface methods (methods in singleton Interfaces)
are still aliased both in their fully parameterized forms and their
unparameterized forms. This works for proxy, but it is not ideal.
Ideally, we would be giving the minimum signature for the interface
method parameters. I just don't have a way of doing that.