A2SEvent reworking; getResource() for inner class
true/false properly not quoted

working on allowing JApplet and Applet to work as focus "windows" as
there is nothing else.

Commented out a check in JTextField that seems unncessary -- only firing
its event if the current focused component is a JTextField. This is a
hack because focus is not yet quite following the mouse