---IMPORTANT CHARACTER SET NOTE---
It is critical that all development work in Java2Script
be done in UTF-8. This means:

- making sure your Exclipse project is set up for UTF-8 (not the Eclipse
default?)
- making sure your server can serve up UTF-8 by default for any
browser-loaded files
- making sure you don't edit a Java2Script class file or one of the site
.js files
  using a non-UTF-8 editor. It may replace non-Latin characters with "?"
or garbage.

In particular, the Mandarin character 秘 (mi; "secret") is used
extensively throughout
the SwingJS class files to distinguish j2s-specific fields and methods
that must not
ever be overridden by subclasses. As in java.lang.Thread.java:

		public static JSThread 秘thisThread;