fix for obscure issue with Interface
 In this case, there is an Interface (TView) that WorldTView implements.
But that interface, in its static initializer, creates a WorldTViewIcon
by referencing a static field in WorldTView,

protected static final Icon WORLDVIEW_ICON =
Tracker.getResourceIcon("axes.gif", true);

Ah, but this is happening before WorldView is fully initialized, and it
causes WorldTView to instantiate all its static fields. OK, but then
later, the JavaScript classloader runs the method that generates all the
defaults for the static fields, and that was nulling out the array that
had just been made.