java2script 5.0.1 Integrated Legacy and SwingJS transpilers
non-SwingJS "Legacy" Jmol was using the java2script 4.2 transpiler which
was only operable using Eclipse Juno (4.2, 2013).

This new branch integrates that into the SwingJS transpiler, allowing
both to be used (and fixed!!) in current Eclipse versions. The
difference is detected by using .j2sjmol for legacy and .j2s for
SwingJS.