java2script dual-mode transpiler 5.0.1 for SwingJS and legacy Jmol
Hooray! Working for Jmol. We can abandon the need to use Eclipse Luna to
run the old java2script 4.2 transpiler that was so quirky.

Now the transpiler for Jmol creates code directly in the site/ folder,
just like the SwingJS transpiler, and changes in "legacy" Jmol can be
checked in JavaScript directly using the ?j2sdebugcode
directive on the url line.